requirements.txt
asgiref==3.8.1
Django==5.0.4
django-rest-passwordreset==1.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
